requirements.txt
chromadb==1.5.0
fastapi==0.129.2
google-genai==1.62.0
openai==2.17.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22
sentence-transformers==5.2.2
torch==2.10.0
transformers==5.1.0
uvicorn==0.40.0
